🤥:: 백준 11382번 꼬마 정민

+++ JavaAlgorithm/src/main/java/org/example/Algorithm/LittleJeongmin.java
@@ -5,9 +5,9 @@ import java.util.Scanner;
 public class LittleJeongmin {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int a = sc.nextInt();
-        int b = sc.nextInt();
-        int c = sc.nextInt();
+        long a = sc.nextInt();
+        long b = sc.nextInt();
+        long c = sc.nextInt();
 
         System.out.println(a+b+c);
     }
